split tests a smaller files
